+++ Tests/Unit/Repository/StatementRepositoryTest.php
@@ -14,8 +14,8 @@ namespace Xabbuh\XApi\Storage\Doctrine\Tests\Unit\Repository;
 use Xabbuh\XApi\DataFixtures\StatementFixtures;
 use Xabbuh\XApi\DataFixtures\VerbFixtures;
 use Xabbuh\XApi\Model\StatementsFilter;
-use Xabbuh\XApi\Storage\Api\Mapping\MappedStatement;
 use Xabbuh\XApi\Storage\Doctrine\Repository\StatementRepository;
+use XApi\Repository\Api\Mapping\MappedStatement;
 
 /**
  * @author Christian Flothmann <christian.flothmann@xabbuh.de>